Move refactor, now supports castle by any side.
Changes in main logic for validations and moves.
Now the move is an object and has it own validation rules
depending on the type of move.
A new module for creating moves is used to detect the type
of move and the move determines if it's valid and it's corresponding
notation.